refactor: split editor plugin workspace helpers
