+++ .sampo/changesets/1000-split-editor-plugin-runtime.md
@@ -1,0 +1,6 @@
+---
+npm/@wp-typia/project-tools: patch
+npm/wp-typia: patch
+---
+
+Split editor-plugin workspace runtime source emitters and bootstrap/registry anchors into focused helper modules.
